Update upload api

+++ src/main/java/com/wuyuncheng/xpress/service/impl/UploadServiceImpl.java
@@ -1,7 +1,8 @@
 package com.wuyuncheng.xpress.service.impl;
 
 import com.wuyuncheng.xpress.config.XPressProperties;
-import com.wuyuncheng.xpress.exception.SaveFileException;
+import com.wuyuncheng.xpress.exception.FileException;
+import com.wuyuncheng.xpress.exception.NotFoundException;
 import com.wuyuncheng.xpress.model.dao.UploadDAO;
 import com.wuyuncheng.xpress.model.dto.UploadDetailDTO;
 import com.wuyuncheng.xpress.model.entity.Upload;
@@ -36,6 +37,16 @@ public class UploadServiceImpl implements UploadService {
 
     @Override
     public void deleteFile(Integer uploadId) {
+        Upload upload = uploadDAO.selectById(uploadId);
+        if (null == upload) {
+            throw new NotFoundException("文件不存在");
+        }
+        String filename = upload.getFilename();
+        File file = new File(properties.getUploadPath() + filename);
+        boolean delete = file.delete();
+        if (!delete) {
+            throw new FileException("文件删除失败");
+        }
         int row = uploadDAO.deleteById(uploadId);
         Assert.state(row != 0, "文件删除失败");
     }
@@ -44,18 +55,18 @@ public class UploadServiceImpl implements UploadService {
     @Override
     public Upload createFile(FileParam fileParam) {
         MultipartFile file = fileParam.getFile();
-        String fileName = DateUtils.nowUnix() + file.getOriginalFilename();
+        String filename = DateUtils.nowUnix() + file.getOriginalFilename();
         // 插入数据库
         Upload upload = new Upload();
         upload.setAuthorId(fileParam.getAuthorId());
-        upload.setFilename(fileName);
+        upload.setFilename(filename);
         upload.setMimetype(file.getContentType());
         upload.setSize(file.getSize());
         upload.setCreated(DateUtils.nowUnix());
         uploadDAO.insert(upload);
         // 写出文件
         if (file.isEmpty()) {
-            throw new SaveFileException("上传文件为空");
+            throw new FileException("上传文件为空");
         }
         try {
             File uploadPath = new File(properties.getUploadPath());
@@ -63,10 +74,10 @@ public class UploadServiceImpl implements UploadService {
                 uploadPath.mkdir();
             }
             byte[] bytes = file.getBytes();
-            Path path = Paths.get(properties.getUploadPath() + fileName);
+            Path path = Paths.get(properties.getUploadPath() + filename);
             Files.write(path, bytes);
         } catch (IOException e) {
-            throw new SaveFileException("文件写入失败");
+            throw new FileException("文件写入失败");
         }
         return upload;
     }
